updated statement  by Developer 1

+++ script.py
@@ -2,7 +2,7 @@
 import math 
 
 # Input: Taking a number from the user 
-num1 = float(input("Enter a number: ")) 
+num1 = int(input("Enter a number: ")) 
 
 # Processing: Calculating the square root 
 # Using math.sqrt from the math module
